requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
async-timeout==4.0.3
attrs==23.2.0
certifi==2024.6.2
charset-normalizer==3.3.2
contourpy==1.2.1
cycler==0.12.1
fonttools==4.53.0
frozenlist==1.4.1
fsspec==2024.6.1
idna==3.7
importlib_resources==6.4.0
Jinja2==3.1.4
joblib==1.4.2
kiwisolver==1.4.5
llvmlite==0.43.0
MarkupSafe==2.1.5
matplotlib==3.8.2
multidict==6.0.5
networkx==3.2.1
numba==0.60.0
numpy==1.26.4
packaging==24.1
pandas==2.2.2
pillow==10.3.0
psutil==6.0.0
pynndescent==0.5.13
pyparsing==3.1.2
python-dateutil==2.9.0.post0
pytz==2024.1
requests==2.32.3
scikit-learn==1.5.0
scipy==1.13.1
six==1.16.0
threadpoolctl==3.5.0
torch==1.10.1+cu113
torch_geometric==2.4.0
tqdm==4.66.4
typing_extensions==4.9.0
tzdata==2024.1
umap-learn==0.5.5
urllib3==2.2.2
yarl==1.9.4
zipp==3.19.2
 